Rental Feature, JWT Authentication

See merge request ankan_debnath/book-rental-microservice!1

+++ user_service/app/api/v1/routes/user_routes.py
@@ -4,36 +4,33 @@ from fastapi import APIRouter
 from fastapi.params import Depends
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette.status import HTTP_201_CREATED, HTTP_204_NO_CONTENT
+
+from app.auth.dependencies import get_current_user
+from app.exceptions.custom_exceptions import CredentialsException
+from app.models import UserModel
 from app.schemas.user import (
     CreateUserRequest,
     UserSchema,
     UpdateUserRequest,
     UpdateUserPatchRequest,
-    Response
+    Response, RentalSchema
 )
 from app.common.db import get_session
 from app.api.v1 import controllers
 
 router = APIRouter(prefix="/user")
 
-@router.post("", status_code=HTTP_201_CREATED)
-async def create_user(
-        user: CreateUserRequest,
-        session:AsyncSession =  Depends(get_session)
-) -> Response:
-
-    new_user = await controllers.add_user(session, user)
-    return Response(
-        success=True,
-        message="User created successfully",
-        data=UserSchema.model_validate(new_user)
-    )
 
 @router.get("/{user_id}")
 async def get_user(
         user_id: uuid.UUID,
+        user: UserModel = Depends(get_current_user),
         session: AsyncSession = Depends(get_session)
 ) -> Response:
+    if str(user.user_id) != str(user_id):
+        raise CredentialsException(
+            detail={"success" : False, "message": "User is unauthorized"}
+        )
     cur_user = await controllers.get_user(session, user_id)
 
     return Response(
@@ -47,9 +44,16 @@ async def get_user(
 async def update_user(
         user_id: uuid.UUID,
         user: UpdateUserRequest,
+        cur_user: UserModel = Depends(get_current_user),
         session: AsyncSession = Depends(get_session)
 )-> Response:
+
+    if str(user_id) != str(cur_user.user_id):
+        raise CredentialsException(
+            detail={"success": False, "message": "User is unauthorized"}
+        )
     updated_user = await controllers.update_user(session, user_id, user)
+
     return Response(
         success=True,
         message="User updated successfully",
@@ -61,8 +65,15 @@ async def update_user(
 async def update_user_patch(
         user_id: uuid.UUID,
         user: UpdateUserPatchRequest,
+        cur_user: UserModel = Depends(get_current_user),
         session: AsyncSession = Depends(get_session)
 ) -> Response :
+
+    if str(user_id) != str(cur_user.user_id):
+        raise CredentialsException(
+            detail={"success": False, "message": "User is unauthorized"}
+        )
+
 
     updated_user = await controllers.update_user_patch(session, user_id, user)
 
@@ -73,11 +84,18 @@ async def update_user_patch(
     )
 
 
-@router.delete("/{user_id}", status_code=HTTP_204_NO_CONTENT)
+@router.delete("/{user_id}")
 async def delete_user(
         user_id: uuid.UUID,
+        cur_user: UserModel = Depends(get_current_user),
         session: AsyncSession = Depends(get_session)
 ) -> Response:
+
+    if str(user_id) != str(cur_user.user_id):
+        raise CredentialsException(
+            detail={"success": False, "message": "User is unauthorized"}
+        )
+
     deleted_user = await controllers.delete_user(session, user_id)
 
     return Response(
@@ -85,7 +103,49 @@ async def delete_user(
         message="User deleted successfully"
     )
 
+@router.post("/{user_id}/rent/{copies}/{book_id}")
+async def rent_book(
+        user_id: uuid.UUID,
+        book_id: uuid.UUID,
+        copies: int,
+        cur_user: UserModel = Depends(get_current_user),
+        session: AsyncSession = Depends(get_session)
+) -> Response:
 
+    if str(user_id) != str(cur_user.user_id):
+        raise CredentialsException(
+            detail={"success": False, "message": "User is unauthorized"}
+        )
+
+    result = await controllers.rent_book(session, user_id, book_id, copies)
+
+    return Response(
+        success=True,
+        message="Book rented successfully",
+        data=RentalSchema.model_validate({"user_id" : user_id, "book_id" : book_id})
+    )
+
+@router.post("/{user_id}/return/{copies}/{book_id}")
+async def rent_book(
+        user_id: uuid.UUID,
+        book_id: uuid.UUID,
+        copies: int,
+        cur_user: UserModel = Depends(get_current_user),
+        session: AsyncSession = Depends(get_session)
+) -> Response:
+
+    if str(user_id) != str(cur_user.user_id):
+        raise CredentialsException(
+            detail={"success": False, "message": "User is unauthorized"}
+        )
+
+    result = await controllers.return_book(session, user_id, book_id, copies)
+
+    return Response(
+        success=True,
+        message="Book returned successfully",
+        data=RentalSchema.model_validate({"user_id" : user_id, "book_id" : book_id})
+    )
